Refactor the dropdown
