--- v2/src/components/dropdownButtons/dropdown.module.css
@@ -1,3 +1,0 @@
-.dropdownContent {
-    z-index:3;
-}
