+++ v2/src/components/header/header.js
@@ -1,9 +1,10 @@
 import React, {useState, useEffect} from "react"
-import logo from '../../assets/images/bitcoin-cash-logo-white-small.png';
+import logo from 'assets/images/bitcoin-cash-logo-white-small.png';
 import { useStaticQuery, graphql } from "gatsby"
 import headerStyles from "./header.module.scss"
-import navBarItems from './navBarItems';
+import Dropdown, { MobileDropdown } from "components/dropdownButtons/dropdown";
 import axios from 'axios';
+import World from 'assets/icons/world.svg';
 import 'assets/lib/hamburgers.min.css';
 import LivePriceWidget from "../liveprice/live-price-widget";
 import Link from 'global/link';
@@ -45,10 +46,45 @@ const externalLink = (i, linkText, href) => {
   )
 }
 
-
-const MobileHeaderLink = ({text, href, id}) => {
-  return <Link className={headerStyles.mobileNavLink} key={id} to={href}>{text}</Link>
+const MobileHeaderLink = ({children, href, id}) => {
+  return <Link className={headerStyles.mobileNavLink} key={id} to={href}>{children}</Link>
 }
+
+const communityDropdownLinks = [
+    {text:"Services", href:"/services.html"},
+    {text:"Projects", href:"/projects.html"},
+    {text:"Exchanges", href:"/exchanges.html"},
+    {text:"Nodes", href:"/nodes.html"},
+    {text:"Developer Portal", href:"/developers.html"},
+    {text: "Whitepaper", href:"/bitcoin.pdf"},
+]
+const languageDropdownLinks = [
+    {text: "English", href:"/"},
+    {text: "Deutsch Deutschland", href:"/de/"},
+    {text: "Español", href:"/es/"},
+    {text: "Français", href:"/fr/"},
+    {text: "日本語", href:"/ja/"},
+    {text: "Nederlands", href:"/nl/"},
+    {text: "Русский", href:"/ru/"},
+    {text: "简体中文", href:"/zh-CN/"},
+    {text: "Español Latin América", href:"/es_419/"},
+    {text: "Bahasa Indonesia", href:"/id/"},
+    {text: "한국어", href:"/ko/"},
+    {text: "Português Brasil", href:"/pt-BR/"},
+    {text: "Türkçe", href:"/tr/"},
+    {text: "繁體中文", href:"/zh-TW/"},
+]
+
+const languageDropdown =(<><World /><span style={{paddingLeft:'5px'}}>Language</span></>);
+
+const navBarItems =[
+    {text:"Get started", href:"/start-here.html"},
+    {text:"Wallets", href:"/wallets.html"},
+    {text:"Logos", href:"/graphics.html"},
+    {text: "Community", links: communityDropdownLinks},
+    {text:"About", href:"/faq.html"},
+    {text: languageDropdown, links: languageDropdownLinks},
+];
 
 const Header = () => {
   const bchPriceApi = "https://min-api.cryptocompare.com/data/price?fsym=BCH&tsyms=USD";
@@ -102,7 +138,7 @@ const Header = () => {
           {navBarItems.map((headerLink, index) =>
             headerLink.href
               ? externalLink(index, headerLink.text, headerLink.href)
-              : headerLink.dropdown
+              : <Dropdown links={headerLink.links} index={index} key={index}>{headerLink.text}</Dropdown>
           )}
         </div>
 
@@ -116,10 +152,12 @@ const Header = () => {
         <div className={headerStyles.mobileMenu} style={isActive ? {height:'calc(100vh - 100px)', backgroundColor: theme.primary_dark} : null}>
           {isActive &&
             <div className={headerStyles.mobileNavLinks}>
-              {navBarItems.map(headerLink =>
+              {navBarItems.map((headerLink, index) =>
                 headerLink.href
-                  ? <MobileHeaderLink text={headerLink.text} href={headerLink.href} key={headerLink.text} id={headerLink.index} />
-                  : headerLink.mobileDropdown
+                  ? <MobileHeaderLink href={headerLink.href} key={headerLink.text} id={headerLink.index}>{headerLink.text}</MobileHeaderLink>
+                  : <MobileDropdown links={headerLink.links} key={index} navLinkClass={headerStyles.mobileNavLink}>
+                      {headerLink.text}
+                    </MobileDropdown>
               )}
             </div>
           }
