+++ v2/src/components/header/header.module.scss
@@ -32,39 +32,38 @@ header {
 }
 
 .headerLinks {
-    a:before, div:before, a:after, div:after {
-        position: absolute;
-        left: 0px;
-        width: 100%;
-        height: 2px;
-        background: rgb(255,141,0);
-        content: "";
-        opacity: 0;
-        transition: all 0.3s;
-    }
+  a:before, div:before, a:after, div:after {
+    position: absolute;
+    left: 0px;
+    width: 100%;
+    height: 2px;
+    background: rgb(255,141,0);
+    content: "";
+    opacity: 0;
+    transition: all 0.3s;
+  }
 
-    a:after, div:after {
-        bottom: 0px;
-        transform: translateY(-10px);
-    }
+  a:after, div:after {
+    bottom: 0px;
+    transform: translateY(-10px);
+  }
 
-    a:hover:after, div:hover:after {
-        opacity: 1;
-        transform: translateY(0px);
-    }
-
-    a, div{
-        position:relative;
-        z-index: 0;
-        cursor: pointer;
-        padding: 5px 0px;
-    }
-}
-
-.topBotomBordersOut 
-{
+  a:hover:after, div:hover:after {
     opacity: 1;
     transform: translateY(0px);
+  }
+
+  a, div {
+    position:relative;
+    z-index: 0;
+    cursor: pointer;
+    padding: 5px 0px;
+  }
+}
+
+.topBotomBordersOut {
+  opacity: 1;
+  transform: translateY(0px);
 }
 
 .link {
@@ -87,7 +86,7 @@ header {
     display:none;
     overflow:scroll;
     /* Hide scrollbar for Chrome, Safari and Opera */
-   &::-webkit-scrollbar {
+    &::-webkit-scrollbar {
         display: none;
     }
 
@@ -103,7 +102,6 @@ header {
     cursor: pointer;
 }
 
-
 .mobileNavLink {
     padding:5px;
     padding-left:15px;
@@ -112,21 +110,20 @@ header {
     display:flex;
 }
 
-
 @media only screen and (max-width: 1100px) {
-    .headerLinks {
-        display:none;
-    }
-    .mobileHeaderLinks {
-        display:block;
-    }
-    .mobileMenu {
-        display:block;
-    }
+  .headerLinks {
+      display:none;
   }
+  .mobileHeaderLinks {
+      display:block;
+  }
+  .mobileMenu {
+      display:block;
+  }
+}
 
-  @media only screen and (max-width: 631px) {
-    .logo {
-        margin-right:14px;
-    }
+@media only screen and (max-width: 631px) {
+  .logo {
+      margin-right:14px;
   }
+}
